ai-service/requirements.txt
# Tomo AI Service — Enterprise Architecture
# Python 3.12+ required

# Framework
fastapi>=0.115,<1.0
sentry-sdk[fastapi]>=2.0,<3.0
uvicorn[standard]>=0.34,<1.0
httpx>=0.28,<1.0
sse-starlette>=2.2,<3.0

# LangGraph + LangChain (let pip resolve compatible versions)
langgraph>=0.4,<1.0
langgraph-checkpoint-postgres>=2.0,<3.0
langchain-anthropic>=0.3,<1.0
langsmith>=0.3,<1.0

# Structured Output
instructor>=1.7,<2.0
pydantic>=2.11,<3.0
pydantic-settings>=2.9,<3.0

# Database
supabase>=2.13,<3.0
asyncpg>=0.30,<1.0
psycopg[binary]>=3.2,<4.0
psycopg-pool>=3.2,<4.0

# Memory (Zep CE — self-hosted)
zep-python>=2.0,<3.0

# RAG (custom PropertyGraphIndex on PostgreSQL — no LlamaIndex runtime dependency)
cohere>=5.13,<6.0

# RAG Quality
ragas>=0.4,<1.0

# Auth
PyJWT>=2.0,<3.0

# Scheduling (LangSmith feedback loop)
apscheduler>=3.10,<4.0

# AI Provider (direct API for fix generation)
anthropic>=0.49,<1.0

# Utilities
python-dotenv>=1.1,<2.0
orjson>=3.10,<4.0
